requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
influxdb==5.3.1
msgpack==1.0.4
paho-mqtt==1.6.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
urllib3==1.26.13